requirements.txt
chris_plugin==0.1.1
pybicpl==0.4.1
